requirements_sd35.txt
torch
diffusers
transformers
accelerate
opencv-python
pillow
protobuf
sentencepiece
huggingface_hub
peft
bitsandbytes
